rename facebook routes too auth-routes
