+++ api/app.js
@@ -13,7 +13,7 @@ var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
 var coursesRouter = require('./routes/courses');
 var reviewsRouter = require('./routes/reviews');
-var facebookRoute = require('./routes/facebookAuth');
+var authRouter = require('./routes/auth-routes');
 
 var app = express();
 
@@ -49,7 +49,7 @@ app.use('/', indexRouter);
 app.use('/users', usersRouter);
 app.use('/courses', coursesRouter);
 app.use('/reviews', reviewsRouter);
-app.use('/auth/facebook', facebookRoute);
+app.use('/auth', authRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
